Move the details method from the News fragment to the StockDetailsFragment
