+++ app/src/main/java/com/lawrence254/stocktrack/adapters/QuotesListAdapter.java
@@ -3,17 +3,20 @@ package com.lawrence254.stocktrack.adapters;
 import android.content.Context;
 import android.content.Intent;
 import android.graphics.Color;
+import android.os.Bundle;
 import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ImageView;
 import android.widget.TextView;
+import android.widget.Toast;
 
 import com.amulyakhare.textdrawable.TextDrawable;
 import com.amulyakhare.textdrawable.util.ColorGenerator;
 import com.lawrence254.stocktrack.R;
 import com.lawrence254.stocktrack.fragments.NewsFragment;
+import com.lawrence254.stocktrack.fragments.StockDetailsFragment;
 import com.lawrence254.stocktrack.model.News;
 import com.lawrence254.stocktrack.model.Quote;
 import com.lawrence254.stocktrack.model.StocksModel;
@@ -69,9 +72,15 @@ public class QuotesListAdapter extends RecyclerView.Adapter<QuotesListAdapter.Qu
             itemView.setOnClickListener(new View.OnClickListener(){
                 @Override
                 public void onClick(View v){
-                    Intent intent = new Intent(v.getContext(), NewsFragment.class);
-                    intent.putExtra("symbol",symbol);
-                    v.getContext().startActivity(intent);
+//                    Intent intent = new Intent(v.getContext(), StockDetailsFragment.class);
+                    Bundle bundle = new Bundle();
+                    bundle.putCharSequence("symbol",symbol);
+
+                    StockDetailsFragment fragment = new StockDetailsFragment();
+                    Toast.makeText(mContext, "Sent Data "+symbol+" to StockDetailsFraagment", Toast.LENGTH_SHORT).show();
+                    fragment.setArguments(bundle);
+//                    intent.putExtra("symbol",symbol);
+//                    v.getContext().startActivity(intent);
 
                 }
             });
